cli: don't fail `nmcli con modify $PROFILE remove $SETTING` for non-existing setting
Removing a setting that is not present should not be an error. The user
asked that the profile doesn't have the requested setting, and that
should succeed (even if that results in no actual change).

Consider when you want to make a hotspot profile "open". That implies
to remove the "wifi-sec" and "802-1x" settings. But you may
not check before whether the profile is already open, and whether
it already has those settings. We should just allow

  $ nmcli connection modify "$PROFILE" remove wifi-sec remove 802-1x

to succeed, regardless whether this changes anything or not.

Likewise, if you do

  $ nmcli connection modify "$PROFILE" con-name foo
  $ nmcli connection modify "$PROFILE" con-name foo

then the second command doesn't fail with "the name is
already \"foo\"". It just succeeds.