cli: replace if-else-if construct in nmc_process_connection_properties() by continue
The if-else-if constuct spans many lines and it is not easy to see that
there is no common action after the if-else-if construct.

Instead, at the end of each if-block, just "continue" the loop. This
is similar to a "return-early" apprach and it mean you don't need
to think what happens at the end of the if-block.